+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [4.2.0](https://github.com/yxx4c/prisma-extension-redis/compare/v4.1.0...v4.2.0) (2026-07-10)
+
+
+### Features
+
+* add includedModels whitelist for auto-caching ([8551bd7](https://github.com/yxx4c/prisma-extension-redis/commit/8551bd7e2662d581dffbb56e8e218ac3d8701a63)), closes [#45](https://github.com/yxx4c/prisma-extension-redis/issues/45)
+
+# [4.1.0](https://github.com/yxx4c/prisma-extension-redis/compare/v4.0.1...v4.1.0) (2026-07-09)
+
+
+### Features
+
+* add direct cache population ([a31f4dc](https://github.com/yxx4c/prisma-extension-redis/commit/a31f4dc9b38333e2b82bd5ebc1565e2842b7a41a))
+* add direct uncache and split exact keys from patterns during invalidation ([e916d18](https://github.com/yxx4c/prisma-extension-redis/commit/e916d1832240e4cd6842274ca7affb38e3ab73f5)), closes [#56](https://github.com/yxx4c/prisma-extension-redis/issues/56)
+
 ## [4.0.1](https://github.com/yxx4c/prisma-extension-redis/compare/v4.0.0...v4.0.1) (2026-07-09)